update snapshot of result of run(config_ex_2)
This test was producing different results on an M1 mac on R versions
greater than 4.1.1 (or 4.1.2?), versus on an intel Mac on R 4.1.1 and in
GitHub actions which used R 4.2.3. It appears to be due to a floating
point error.

Specifically, the number of days that a duration-based transition
lasts is determined by the first day that the cumulative sum of the
daily transition probabilities > 1. This number was being calculated
differently on the different systems - specifically for the transition
to the life stage '__n'.

This test is brittle, but not sure what a longer term fix would look
like.