attempting to make winter tick stable
The version of the winter tick config that was adapted to the newest config structure had a declining population in the resulting model output. Upon comparing the config to the version used in `dave_script.R` in the state of repo at commit `40b8218717f51481033f801033346ee93595e6f6`, it seems that several components of the config are semantically different.

With the changes in this commit, I'm again getting a fairly stable model output, using the newest config structure. There are still differences between the current output and the one from the prior commit I referenced, but I'm not sure yet whether these are meaningful differences.